a new fleet control with only one model; flee control

+++ docker/disable.sh
@@ -1,5 +1,13 @@
-sudo docker stop jetbot_camera 
-sudo docker rm jetbot_camera
+
+source configure.sh
+
+JUPYTER_WORKSPACE=${1:-$HOME}  # default to $HOME
+JETBOT_CAMERA=${2:-opencv_gst_camera}  # default to opencv
+
+if [ "$JETBOT_CAMERA" = "zmq_camera" ]; then
+  sudo docker stop jetbot_camera
+  sudo docker rm jetbot_camera
+fi
 
 sudo docker stop jetbot_jupyter 
 sudo docker rm jetbot_jupyter
